+++ CHANGELOG.md
@@ -1,1 +1,29 @@
 # Change Log
+
+## 1.0.0 - 2023-03-22
+
+### vscode engines
+
+- 1.76.2
+
+### Added
+
+- insertString option
+- isIncludeId option
+- excludeRegex option
+
+### Changed
+
+- None
+
+### Deprecated
+
+- None
+
+### Removed
+
+- Remove pop-up insertion of characters by InputBox
+
+### Fixed
+
+- None